+++ docs/jquery.loft-labels.js
@@ -1,5 +1,5 @@
 /**
- * Loft Labels jQuery Plugin v1.1.0
+ * Loft Labels jQuery Plugin v1.1.1
  * http://www.intheloftstudios.com/packages/js/jquery.loft_labels
  *
  * jQuery plugin to move labels into the input element as placeholders with optional lightweight input validation.
@@ -7,7 +7,7 @@
  * Copyright 2013-2018,
  * @license MIT
  *
- * Date: Sat Dec 15 08:37:31 PST 2018
+ * Date: Sun Dec 16 09:03:57 PST 2018
  */
 (function(factory) {
   if (typeof define === 'function' && define.amd) {
@@ -63,7 +63,8 @@
           $validation: $members,
           defaultText: null,
           settings: settings,
-          hasPlaceholderText: null,
+          segment: {},
+          hasDefaultText: true,
           states: [],
 
           /**
@@ -101,6 +102,10 @@
               var name = 'data-' + this.settings.dataPrefix + this.segment.name,
                 placeholder = this.$label.attr(name) || placeholder;
             }
+            if (typeof this.settings.onGetLabel === 'function') {
+              placeholder = this.settings.onGetLabel.call(this, placeholder);
+            }
+
             return placeholder;
           },
 
@@ -115,14 +120,14 @@
 
         $el.data(pluginName, instance);
         initializeInstance.call(instance);
-        detectDefaultState(instance);
+        // detectDefaultState(instance);
 
         // Handlers
         $el
           .bind('focus', function() {
-            instance.hasPlaceholderText || detectDefaultState(instance);
+            instance.hasDefaultText || detectDefaultState(instance);
             instance.states.focus = true;
-            if (instance.hasPlaceholderText) {
+            if (instance.hasDefaultText) {
               $el.val('');
               renderHtml.call(instance);
             }
@@ -156,22 +161,20 @@
       });
 
     if (this.settings.breakpointX) {
-      this.settings.breakpointX
-        .addCrossAction(function(segment) {
-          // Read in the segment label for all elements.
-          $elements.each(function() {
-            var instance = $(this).data('loftLabels');
-            instance.segment = segment;
-            refreshInstance(instance);
-          });
-        })
-        .triggerActions();
+      this.settings.breakpointX.addCrossAction(function(segment) {
+        // Read in the segment label for all elements.
+        $elements.each(function() {
+          var instance = $(this).data('loftLabels');
+          instance.segment = segment;
+          refreshInstance(instance);
+        });
+      });
     }
 
     return this;
   };
   $.fn.loftLabels.version = function() {
-    return '1.1.0';
+    return '1.1.1';
   };
   $.fn.loftLabels.defaults = {
     // The class to add to the textfield when it's in focus.
@@ -206,16 +209,23 @@
     breakpointX: null,
 
     /**
+     * Preset the default segement instead of using window.width.
+     *
+     * By default the segment at instantiation is based on the window width.
+     * You may pass a segment here to use instead.  This may only be necessary
+     * for unti testing.
+     */
+    segment: null,
+
+    /**
      * Modify the label string.
      *
-     * Use this to alter the label string before it's placed into the
-     * textfield. The instance is available as 'this'.  This happens once
-     * during the init phase.
+     * Use this to alter the label string before getLabel().
      *
      * @param defaultText
      * @returns {*}
      */
-    callback: null,
+    onGetLabel: null,
 
     /**
      * A function to call after init has completed.
@@ -314,7 +324,7 @@
   };
 
   function detectDefaultState(instance) {
-    instance.hasPlaceholderText = instance.isDefault() || !instance.getValue();
+    instance.hasDefaultText = instance.isDefault() || !instance.getValue();
   }
 
   function initializeInstance() {
@@ -333,25 +343,24 @@
 
     // Modify the default text...
     this.defaultText = $.trim(this.defaultText);
-    this.defaultText = s.callback
-      ? s.callback.call(this, this.defaultText)
-      : this.defaultText;
 
     // If we have a value in the form, this plugin is moot.
-    if (this.getValue()) {
-      this.$el.addClass(s.focus);
-      return $(this);
-    } else {
-      this.setValue(this.getLabel());
+    this.hasDefaultText = false;
+    if (!this.getValue()) {
+      this.hasDefaultText = true;
+      if (s.breakpointX) {
+        this.segment = s.segment || s.breakpointX.getSegmentByWindow();
+      }
+      // Due to complex reasons; do not use setValue() here! 2018-12-15T13:27,
+      // aklump
+      this.$el.val(this.getLabel());
       this.$el.removeClass(s.focus);
     }
-
     renderHtml.call(this);
     validationHandler(this, { type: 'init' }, this.$validation);
     if (typeof s.onInit === 'function') {
       s.onInit(this);
     }
-    detectDefaultState(this);
 
     return this;
   }
@@ -398,7 +407,7 @@
    * If displaying default text, make sure it's correct.
    */
   function refreshInstance(instance) {
-    if (instance.hasPlaceholderText) {
+    if (instance.hasDefaultText) {
       instance.$el.val(instance.getLabel());
       renderHtml.call(instance);
       return this;